+++ src/matrix/ft/CGRAPPA.hpp
@@ -139,7 +139,7 @@ public:
         
 		Matrix<CT> res = kspace;
 #pragma omp parallel for default (shared)
-		for (int coil = 0; coil < m_nc; ++coil)
+		for (int coil = 0; coil < (int)m_nc; ++coil)
 			Slice(res, coil, ARC(kspace,coil));
 		return res;
 	}
